+++ src/idtrackerai/tracker/identity_transfer.py
@@ -4,7 +4,7 @@ import numpy as np
 from torch.nn import Module
 
 from idtrackerai import GlobalFragment, Video
-from idtrackerai.network import fc_weights_reinit
+from idtrackerai.network import NetworkParams, fc_weights_reinit
 from idtrackerai.utils import conf
 
 from .accumulation_manager_utils import (
@@ -14,7 +14,6 @@ from .accumulation_manager_utils import (
 )
 from .assigner import compute_identification_statistics_for_non_accumulated_fragments
 from .network.get_predictions import GetPredictionsIdentities
-from .network.network_params import NetworkParams
 
 
 def identify_first_global_fragment_for_accumulation(